% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{combatCorrect}
\alias{combatCorrect}
\title{combatCorrect - It corrects the batch effect from the gene expression matrix using comBat function}
\usage{
combatCorrect(
  preparedExprData,
  preparedCovs,
  batchCov,
  bioCovsToCorrect,
  PCA_plots = F,
  name = deparse(substitute(preparedExprData)),
  outputDir = paste0(getwd(), "pcaPlots/")
)
}
\arguments{
\item{preparedExprData}{the gene expression matrix of a specific cell type cluster obtained from prepareData function}

\item{preparedCovs}{the metadata table of a specific cell type cluster obtained from prepareData function}

\item{batchCov}{the name of the variable that represents the batch}

\item{bioCovsToCorrect}{other covariates besides batch}

\item{PCA_plots}{if true, PCA plots will be saved}

\item{name}{the name of the cell type cluster (i.e.microglia_cluster1)}

\item{outputDir}{if save=TRUE, in which directory the plot will be saved}
}
\value{
it returns the gene expression matrix corrected by batch effect
}
\description{
combatCorrect - It corrects the batch effect from the gene expression matrix using comBat function
}
