% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{get_all_GCNs}
\alias{get_all_GCNs}
\title{get_all_GCNs - It creates and annotates a set of GCNs for a specific cell type cluster}
\usage{
get_all_GCNs(
  exprData,
  metadata,
  gcnName = "GCN1",
  idColname = NULL,
  batchColname = NULL,
  minNumCells = 200,
  covsToCorrect = NULL,
  filteringGenes = F,
  filteringCutoff = NULL,
  filteringProportion = NULL,
  plots = F,
  path = getwd(),
  moduleTraitCorr = F,
  covsToCorr = NULL,
  phenotypeEnrich = T,
  correction_method = "gSCS",
  sources = c("GO", "KEGG", "REAC"),
  organism = "hsapiens",
  exclude_iea = T
)
}
\arguments{
\item{exprData}{a normalized cell type cluster gene expression matrix}

\item{metadata}{a cell type cluster metadata table}

\item{gcnName}{the name of the network}

\item{idColname}{the name of the column that contains donors' ID}

\item{batchColname}{the name of the column that contains batch effect information}

\item{minNumCells}{the minimum number of cells to create a GCN}

\item{covsToCorrect}{the biological or technical covariates names that we are not interested in and we want to remove their effect on our data}

\item{filteringGenes}{if filteringGenes=TRUE, we want to apply gene filtering based on a minimum expression value or cutoff on a percentage of cells}

\item{filteringCutoff}{if filteringGenes=TRUE, filteringCutoff represents the minimum expression value}

\item{filteringProportion}{if filteringGenes=TRUE, filteringProportion represents the percentage of cells}

\item{plots}{if plots=TRUE, plots will be saved}

\item{path}{the path where the files will be saved}

\item{moduleTraitCorr, }{if moduleTraitCorr=T, the asociation between module eigengenes and traits will be estimated}

\item{sources}{the databases we want to use for functional enrichment analysis (default databases are GO, KEGG and REAC)}

\item{organism}{the specie of the organism}

\item{covsToCorr=NULL, }{if moduleTraitCorr=T, covsToCorr includes the traits to be evaluated}

\item{phenotypeEnrich=T, }{if phenotypeEnrich=T, phenotype enrichment analysis will be carried out for each module of the network}

\item{correction.method}{the correction method used for functional enrichment analysis (deafult method is gSCS)}

\item{exclude.iea}{if exclude.iea=TRUE, we remove automatically annotations (not manually curated)}
}
\value{
it returns a new folder containing theses files: the residuals, the network, the functional enrichment analysis results,
the phenotype enrichment analysis results, the cell type markers enrichment and module membership.
}
\description{
get_all_GCNs - It creates and annotates a set of GCNs for a specific cell type cluster
}
