% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{getColors}
\alias{getColors}
\title{getColors - It returns the colors to be used for the sunburst plot}
\usage{
getColors(
  criteria = c("T0_modules", "all_modules", "covariate"),
  modulesGenes = NULL,
  nets_dir = NULL,
  metadata = NULL,
  cellID = "cell_id",
  covariate = c("Clinical_Dx")
)
}
\arguments{
\item{criteria}{the criteria used to color the modules: "T0_modules", "all_modules" or "covariate".}

\item{modulesGenes}{if criteria!="covariate", a data frame with one row per gene and one column per iteration is needed. Each}

\item{nets_dir}{if criteria=="covariate", the path with all the networks that belong to the same cell type and different iterations}

\item{metadata}{if criteria=="covariate", the path or the data frame with the metadata of all the cells that belong to a specific cell type}

\item{cellID}{if criteria=="covariate", the name of the variable that contains the ID of each cell in the metadata table}

\item{covariate}{if criteria=="covariate", the name of the biological covariate we are interested in}
}
\value{
a list containing both the colors assigned and the names of the modules
}
\description{
getColors - It returns the colors to be used for the sunburst plot
}
