% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{getEnrichmentPlot}
\alias{getEnrichmentPlot}
\title{getEnrichmentPlot - It returns an enrichment plot for a T0 module of interest and show how the enrichment changes through pseudo-cells iterations
for this same set of genes.}
\usage{
getEnrichmentPlot(
  module_name = "pink",
  source = "GO:BP",
  modulesGenes = NULL,
  nets_dir = NULL
)
}
\arguments{
\item{source}{the name of the source to be used for the enrichment analysis (GO, KEGG, REAC).}

\item{modulesGenes}{a data frame with as many rows as genes and as many columns as iterations. For each gene (row), you can see the module where
it belongs in each pseudo-cell iteration.}

\item{nets_dir}{if modulesGenes is not provided, the path where all the networks of the same cell type and different iterations are located.}
}
\value{
an enrichment plot
}
\description{
getEnrichmentPlot - It returns an enrichment plot for a T0 module of interest and show how the enrichment changes through pseudo-cells iterations
for this same set of genes.
}
