% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_annotation.R
\name{getFunctionalEnrichment}
\alias{getFunctionalEnrichment}
\title{getFunctionalEnrichment - It returns the results of a functional enrichment analysis applied on each module of the selected network.}
\usage{
getFunctionalEnrichment(
  net,
  correction_method = "gSCS",
  sources = c("GO", "KEGG", "REAC"),
  organism = "hsapiens",
  exclude_iea = T
)
}
\arguments{
\item{net}{the pathway of the network selected for the functional enrichment analysis}

\item{sources}{the databases we want to use for functional enrichment analysis (default databases are GO, KEGG and REAC)}

\item{organism}{the specie of the organism}

\item{correction.method}{the correction method used for functional enrichment analysis (deafult method is gSCS)}

\item{exclude.iea}{if exclude.iea=TRUE, we remove automatically annotations (not manually curated)}
}
\value{
it returns the data frame containing the results of gprofiler2, where each row represents
one annotation associated with a specific module
}
\description{
getFunctionalEnrichment - It returns the results of a functional enrichment analysis applied on each module of the selected network.
}
