% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{getInteractiveBrowser}
\alias{getInteractiveBrowser}
\title{getInteractiveBrowser - It creates an interactive browser with T0 modules annotations and highlights the most relevant modules based on user's criteria.}
\usage{
getInteractiveBrowser(nets.dir, path, name, criteria = NULL)
}
\arguments{
\item{nets.dir}{the path where the networks of the same cell type and different iterations are located.}

\item{path}{the directory where the statistics will be saved.}

\item{name}{the name of the cell type.}

\item{criteria}{the criteria to highlight the most relevant modules.}
}
\value{
it returns an interative table with one row for each T0 module and some columns that shows the main annotations of each module.
}
\description{
getInteractiveBrowser - It creates an interactive browser with T0 modules annotations and highlights the most relevant modules based on user's criteria.
}
