% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_annotation.R
\name{getMarkers}
\alias{getMarkers}
\title{getMarkersList - It returns a list of markers.}
\usage{
getMarkers(
  type = c("cellTypeMarkers", "dopaminergicMarkers", "ferroptosisMarkers", "PDgenes")
)
}
\arguments{
\item{type}{it indicates the kind of markers you want to get}
}
\value{
the list of markers selected
}
\description{
getMarkersList - It returns a list of markers.
}
