% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_annotation.R
\name{getMarkersListEnrich}
\alias{getMarkersListEnrich}
\title{getMarkersListEnrich - It applies a gene set enrichment analysis on each module of the network for a list of markers.}
\usage{
getMarkersListEnrich(net, markersList = "cellTypeMarkers", exprData)
}
\arguments{
\item{net}{the name of the network}

\item{markersList}{the list of markers to be used for the enrichment. It can be one of the list markers available in scCoExpNets or a list of markers provided by the user.}

\item{exprData}{the gene expression matrix used to create the network}
}
\value{
a data frame containing the results of the GSEA. We will find one row per module-markers.
}
\description{
getMarkersListEnrich - It applies a gene set enrichment analysis on each module of the network for a list of markers.
}
