% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{getModules}
\alias{getModules}
\title{getModules - It creates and annotates a single-cell gene co-expression network
It is based on getDownstreamNetwork function from CoExpNets.}
\usage{
getModules(
  tissue = "mytissue",
  n.iterations = 50,
  min.exchanged.genes = 20,
  expr.data,
  beta = -1,
  job.path = "~/tmp/",
  min.cluster.size = 30,
  net.type = "signed",
  debug = F,
  blockTOM = F,
  save.tom = F,
  save.plots = F,
  excludeGrey = FALSE,
  fullAnnotation = F,
  silent = T
)
}
\arguments{
\item{tissue}{A label to use to refer to the results in files and downstream results}

\item{n.iterations}{Number of iterations in the k-means algorithm to refine the
clustering process}

\item{expr.data}{A data frame with cell type cluster expression data obtained from svaCorrection function (genes at cols and cells at rows)}

\item{beta}{The smoothing parameter of the WGCNA pipeline. If -1, then the method
will suggest one automatically by looking at the R2 between gene connectivity and
Scale Free Topology features}

\item{job.path}{This method will generate a number of files. It needs a folder to write
results}

\item{min.cluster.size}{Minimum number of genes for a group to be considered as cluster
for the tree cutting algorithm to convert from a dendogram to a cluster.}

\item{net.type}{Whether a signed ("signed") or unsigned ("unsigned") network type will be created}

\item{debug}{Set this to true if you want a quick run of the method to test how it works with
a small amount of your genes}

\item{excludeGrey}{If WGCNA detects grey genes, set it to TRUE if you want them removed
from the network before applying k-means}

\item{fullAnnotation}{Set to TRUE if you want to annotate by cell type and GO, REACTOME and KEGG}

\item{fastKMeans}{Set to TRUE for a faster k-means use}
}
\value{
A file name that can be used to access your network

it returns a data frame with one column per module, one row per cell type markers list and each cell containing
the result of a Fisher's exacta test.
}
\description{
getModules - It creates and annotates a single-cell gene co-expression network
It is based on getDownstreamNetwork function from CoExpNets.
}
