% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_annotation.R
\name{getPhenotypeTerms}
\alias{getPhenotypeTerms}
\title{getPhenotypeTerms - It returns the results of a phenotype enrichment analysis applied on each module of the selected network.}
\usage{
getPhenotypeTerms(net)
}
\arguments{
\item{net.file}{the pathway of the network selected for the functional enrichment analysis}
}
\value{
it returns the data frame containing the results of the phenotype enrichment analysis, where each row represents
one annotation associated with a specific module
}
\description{
getPhenotypeTerms - It returns the results of a phenotype enrichment analysis applied on each module of the selected network.
}
