% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{getPreservedModules}
\alias{getPreservedModules}
\title{getPreservedModules  - Check whether one network module is preserved in a tissue gene expression profile}
\usage{
getPreservedModules(
  network,
  expr.data.files = NULL,
  tissues = c("snig", "putm"),
  permutations = 200,
  maxModuleSize = 5000,
  maxGoldModuleSize = 400,
  randomSeed = 1
)
}
\arguments{
\item{network}{The network you want to assess about preservation (full path and file name or a network object).}

\item{expr.data.files}{A list with the expression data used to create the network (1st element) and the
expression data onto which we want the network assessed of preservation (2nd element)}

\item{tissues}{A vector of strings, to name the tissues}

\item{permutations}{The number of random permutations used to construct the null hypothesis. 200 may be a
reasonable number}

\item{maxModuleSize}{All modules over this size in genes will be considered as having as much size as this limit}

\item{maxGoldModuleSize}{The size of the random module to be used as a reference of a totally random module values}

\item{randomSeed}{Seed for the random number generator}
}
\value{
A data frame with modules in the rows and statistics at the columns.
}
\description{
getPreservedModules  - Check whether one network module is preserved in a tissue gene expression profile
}
