% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{getRobustAndEvolvedModules}
\alias{getRobustAndEvolvedModules}
\title{getRobustAndEvolvedModules - For each module of the T0 network, it returns the percentage of the genes maintained together at Tn.}
\usage{
getRobustAndEvolvedModules(nets.dir, path = getwd(), name = "CT1")
}
\arguments{
\item{nets.dir}{the path where the networks of the same cell type and different iterations are located.}

\item{path}{the directory where the statistics will be saved.}

\item{name}{the name of the cell type.}
}
\value{
it returns a table with one row for each T0 module and some columns that show how the module composition has changed from T0 to Tn.
}
\description{
getRobustAndEvolvedModules - For each module of the T0 network, it returns the percentage of the genes maintained together at Tn.
}
