% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{getSankeyPlot}
\alias{getSankeyPlot}
\title{getSankeyPlot - It returns a sankey plot that shows how the genes of a T0 module of interest are distributed in other modules through pseudo-cells
iterations. The plot contains as many columns as pseudo-cells iterations and each column contains as many boxes as modules in that iteration.}
\usage{
getSankeyPlot(module_name = "pink", nets_dir = NULL)
}
\arguments{
\item{module_name}{the name of the T0 module selected to get the sankey plot.}

\item{nets_dir}{if modulesGenes is not provided, the path where all the networks of the same cell type and different iterations are located.}
}
\value{
a sankey plot
}
\description{
getSankeyPlot - It returns a sankey plot that shows how the genes of a T0 module of interest are distributed in other modules through pseudo-cells
iterations. The plot contains as many columns as pseudo-cells iterations and each column contains as many boxes as modules in that iteration.
}
