% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{getSunburstPlot}
\alias{getSunburstPlot}
\title{getSunburstPlot - It returns a sunburst plot where the inner ring represent the T0 modules (the size of each segment is proportional to the number
of genes) and the rest of rings represent how the genes of each T0 module remains or not together after successive pseudo-cells iterations}
\usage{
getSunburstPlot(
  criteria = c("T0_modules", "all_modules", "covariate"),
  nets_dir = NULL,
  metadata = NULL,
  cellID = "cell_id",
  covariate = c("Clinical_Dx"),
  showLabels = F
)
}
\arguments{
\item{criteria}{the criteria used to color the modules: "T0_modules", "all_modules" or "covariate".}

\item{nets_dir}{the path with all the networks that belong to the same cell type and different iterations.}

\item{metadata}{if criteria=="covariate", the path or the data frame with the metadata of all the cells that belong to a specific cell type}

\item{cellID}{if criteria=="covariate", the name of the variable that contains the ID of each cell in the metadata table}

\item{covariate}{if criteria=="covariate", the name of the biological covariate we are interested in}
}
\value{
a list containing both the colors assigned and the names of the modules
}
\description{
getSunburstPlot - It returns a sunburst plot where the inner ring represent the T0 modules (the size of each segment is proportional to the number
of genes) and the rest of rings represent how the genes of each T0 module remains or not together after successive pseudo-cells iterations
}
