% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_annotation.R
\name{getTraitCorr}
\alias{getTraitCorr}
\title{getTraitCorr - It applies a linear regression approach to evaluate the association between the eigengenes of one module and donors' covariates}
\usage{
getTraitCorr(net, metadata, covsToCorr = NULL)
}
\arguments{
\item{net}{the network with the modules to be annotated}

\item{metadata}{the data frame containing the covariates information for each cell}

\item{covsToCorr}{covariates to evaluate}
}
\value{
a data frame containing the results of the module-trait association. This data frame contains as many rows as covariates.
}
\description{
getTraitCorr - It applies a linear regression approach to evaluate the association between the eigengenes of one module and donors' covariates
}
