% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{pair}
\alias{pair}
\title{pair - It combines single-cell metadata and corresponding individual covariates in the same data frame
It is an auxiliary function for pseudoCells function}
\usage{
pair(indexes, paired)
}
\arguments{
\item{indexes}{it represents a range of numbers from 1 to the number of columns of the gene expression matrix}

\item{pair}{it represents the ID of the cells. To group the cells per individual, it is necessary that individual ID is included in cell ID}
}
\value{
a data frame containing pairs of cells grouped per individual ID. The data frame contains three columns: cell1, cell2, sampleID
}
\description{
pair - It combines single-cell metadata and corresponding individual covariates in the same data frame
It is an auxiliary function for pseudoCells function
}
