% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{plotMDS_Batch}
\alias{plotMDS_Batch}
\title{plotMDS_Batch - It plots the cells based on their expression using a MDS plot in order to visualize a possible batch effect}
\usage{
plotMDS_Batch(exprData, covs, batchCov, name, save = F, outputDir = getwd())
}
\arguments{
\item{exprData}{a normalized cell type cluster matrix}

\item{covs}{a cell type cluster metadata table}

\item{batchCov}{the name of the variable that represents the batch}

\item{name}{the name of the network}

\item{save}{if save=TRUE, the plot will be saved}

\item{outputDir}{if save=TRUE, in which directory the plot will be saved}
}
\value{
it shows a MDs plot with a small set of cells
}
\description{
plotMDS_Batch - It plots the cells based on their expression using a MDS plot in order to visualize a possible batch effect
}
