% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{prepareData}
\alias{prepareData}
\title{prepareData - It checks if the gene expression matrix and the metadata files are ready for batch effect correction and/or surrogate variable analysis.}
\usage{
prepareData(
  exprData,
  covs,
  batchCov = NULL,
  idCov,
  filteringGenes = F,
  filteringCutoff = NULL,
  filteringProportion = NULL
)
}
\arguments{
\item{exprData}{a normalized cell type cluster matrix}

\item{covs}{a cell type cluster metadata table}

\item{batchCov}{the name of the variable that represents the batch}

\item{idCov}{the name of the variable that represents the individuals ID}

\item{filteringGenes}{if true, genes are filtered based on their minimum expression (filteringCutoff) in a percentage of cells (filteringProportion)}

\item{filteringProportion}{if filteringGenes=TRUE, filteringProportion represents the proportion of cells}

\item{filteringCuroff}{if filteringGenes=TRUE, filteringCutoff represents the minimum value expression}
}
\value{
it returns a list containing our data ready. The first element represents the gene expression matrix and the second one, the metadata table.
}
\description{
prepareData - It checks if the gene expression matrix and the metadata files are ready for batch effect correction and/or surrogate variable analysis.
}
