% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_visualization.R
\name{summariseModulesAnnotations}
\alias{summariseModulesAnnotations}
\title{summariseModulesAnnotations - For each module of the T0 network, it returns the main annotations, including the size, the percentage of genes maintained together at Tn, the association with clinical covariates and the association with different list of markers.}
\usage{
summariseModulesAnnotations(nets.dir = nets.dir, path = getwd(), name = "CT1")
}
\arguments{
\item{nets.dir}{the path where the networks of the same cell type and different iterations are located.}

\item{path}{the directory where the statistics will be saved.}

\item{name}{the name of the cell type.}
}
\value{
it returns a table with one row for each T0 module and some columns that shows tha main annotations of each module.
}
\description{
summariseModulesAnnotations - For each module of the T0 network, it returns the main annotations, including the size, the percentage of genes maintained together at Tn, the association with clinical covariates and the association with different list of markers.
}
