% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGCNs_creation.R
\name{svaCorrection}
\alias{svaCorrection}
\title{svaCorrection - It applies a surrogate variable analysis in order to remove biological and/or technical effects that we are not interested in}
\usage{
svaCorrection(
  combatExprData,
  preparedCovs,
  bioCovsToCorrect,
  idCov,
  batchCov = NULL,
  plots = T,
  outputDir = paste0(getwd(), "/svas_plots/"),
  name = deparse(substitute(combatExprData)),
  save = T,
  residsOutput = paste0(getwd(), "/residuals/")
)
}
\arguments{
\item{combatExprData}{the gene expression matrix of a specific cell type cluster obtained from combatCorrect function}

\item{preparedCovs}{the metadata table of a specific cell type cluster obtained from prepareData function}

\item{bioCovsToCorrect}{the covariates (biological, technical or both) that we are not interested in and we want to remove their effect on our dataset}

\item{idCov}{the name of the variable that represents the individual ID}

\item{batchCov}{the name of the variable that represents the batch}

\item{plots}{if plots=TRUE, plots will be saved}

\item{outputDir}{if save=TRUE, in which directory the plot will be saved}

\item{name}{the name of the cell type cluster (i.e.microglia_cluster1)}

\item{save}{if save=TRUE, the residuals file will be saved}

\item{residsOutout}{if save=TRUE, the directory where the residuals file will be saved}
}
\value{
it returns the residuals matrix (genes at columns and cells at rows)
}
\description{
svaCorrection - It applies a surrogate variable analysis in order to remove biological and/or technical effects that we are not interested in
}
